Added validation in CRUD operations

+++ Project/src/main/java/project/services/implementation/CategoryServiceImplementation.java
@@ -2,6 +2,8 @@ package project.services.implementation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import project.exceptions.EntityAlreadyExistsException;
+import project.exceptions.EntityNotFoundException;
 import project.models.CategoryModel;
 import project.repositories.CategoryRepository;
 import project.services.CategoryService;
@@ -15,27 +17,36 @@ public class CategoryServiceImplementation implements CategoryService {
 
     @Override
     public CategoryModel create(CategoryModel category) {
-        CategoryModel toReturn = categoryRepository.save(category);
-        return toReturn;
+        if (categoryRepository.findByName(category.getName()) != null)
+            throw new EntityAlreadyExistsException("Category", "name", category.getName());
+        return categoryRepository.save(category);
     }
 
     @Override
     public CategoryModel read(Integer id) {
         CategoryModel toReturn = categoryRepository.findById(id);
+        if (toReturn == null)
+            throw new EntityNotFoundException("Category", id);
         return toReturn;
     }
 
     @Override
     public CategoryModel update(Integer id, CategoryModel category) {
         CategoryModel toReturn = categoryRepository.findById(id);
+        if (toReturn == null)
+            throw new EntityNotFoundException("Category", id);
+        if (categoryRepository.findByName(category.getName()) != null)
+            throw new EntityAlreadyExistsException("Category", "name", category.getName());
         if (category.getName() != null) toReturn.setName(category.getName());
         if (category.getDescription() != null) toReturn.setDescription(category.getDescription());
-        categoryRepository.save(toReturn);
-        return toReturn;
+        return categoryRepository.save(toReturn);
     }
 
     @Override
     public void delete(Integer id) {
+        CategoryModel toReturn = categoryRepository.findById(id);
+        if (toReturn == null)
+            throw new EntityNotFoundException("Category", id);
         categoryRepository.delete(id);
     }
