+++ Project/src/main/java/project/services/implementation/PropertyServiceImplementation.java
@@ -2,6 +2,7 @@ package project.services.implementation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import project.exceptions.EntityNotFoundException;
 import project.models.CategoryModel;
 import project.models.PropertyModel;
 import project.repositories.CategoryRepository;
@@ -19,30 +20,35 @@ public class PropertyServiceImplementation implements PropertyService {
 
     @Override
     public PropertyModel create(PropertyModel property) {
-        PropertyModel toReturn = propertyRepository.save(property);
-        return toReturn;
+        return propertyRepository.save(property);
     }
 
     @Override
     public PropertyModel read(Integer id) {
         PropertyModel toReturn = propertyRepository.findById(id);
+        if (toReturn == null)
+            throw new EntityNotFoundException("Property", id);
         return toReturn;
     }
 
     @Override
     public PropertyModel update(Integer id, PropertyModel property) {
         PropertyModel toReturn = propertyRepository.findById(id);
+        if (toReturn == null)
+            throw new EntityNotFoundException("Property", id);
         if (property.getDescription() != null) toReturn.setDescription(property.getDescription());
         if (property.getRooms() != null) toReturn.setRooms(property.getRooms());
         if (property.getSize() != null) toReturn.setSize(property.getSize());
         if (property.getState() != null) toReturn.setState(property.getState());
         if (property.getOperation() != null) toReturn.setOperation(property.getOperation());
-        propertyRepository.save(toReturn);
-        return toReturn;
+        return propertyRepository.save(toReturn);
     }
 
     @Override
     public void delete(Integer id) {
+        PropertyModel toReturn = propertyRepository.findById(id);
+        if (toReturn == null)
+            throw new EntityNotFoundException("Property", id);
         propertyRepository.delete(id);
     }
